# optimizing is_ipv4