--- a/.github/workflows/visual_studio.yml
+++ b/.github/workflows/visual_studio.yml
@@ -42,4 +42,4 @@ jobs:
       run: cmake --build build --config "${{matrix.config}}" --verbose
     - name: Run  tests
       working-directory: build
-      run: ctest -C Debug  --output-on-failure
+      run: ctest -C "${{matrix.config}}"  --output-on-failure