@anonrig 2023-11-13
I think we have a bug in here. This is somewhat similar to https://github.com/nodejs/node/issues/50632

```
D:\a\ada\ada\tests\wpt_tests.cpp(385): error: Expected equality of these values:
  input_url->get_host()
    Which is: "256"
  host
    Which is: "0.0.1.0"
```