@lemire 2023-11-13
@anonrig Seems unrelated to this Node issue unless I misunderstand something. I have a Windows box open in front of me and I am failing to reproduce this issue.