@lemire 2023-11-13
The failing test is this...

```json
{
    "input": "http://256",
    "base": "http://other.com/",
    "href": "http://0.0.1.0/",
    "origin": "http://0.0.1.0",
    "protocol": "http:",
    "username": "",
    "password": "",
    "host": "0.0.1.0",
    "hostname": "0.0.1.0",
    "port": "",
    "pathname": "/",
    "search": "",
    "hash": ""
  }
```

But this test is not new. And I cannot reproduce it locally.

And it is only triggered by Visual Studio in release mode. And only in the last few commits.

Weird. 

What could it be?