@lemire 2023-11-13
@anonrig I am merging this. We can revert later if you disagree.